Update log for system

+++ src/main/java/com/khiemtran/service/impl/AuthenticationServiceImpl.java
@@ -53,7 +53,8 @@ public class AuthenticationServiceImpl implements AuthenticationService {
   @Override
   public UserResponse register(SignUpRequest request) {
     if (userRepository.existsByEmail(request.email())) {
-      throw new EmailNotFoundException("User has already exist.");
+      log.error("Registration attempt failed: User with email {} already exists.", request.email());
+      throw new EmailNotFoundException("User already exists.");
     }
     Role roles = getRoles();
     User userEntity = new User(
@@ -64,6 +65,7 @@ public class AuthenticationServiceImpl implements AuthenticationService {
         request.city(),
         Collections.singleton(roles));
     User user = userRepository.save(userEntity);
+    log.info("User successfully registered with username: {} and email: {}", user.getUsername(), user.getEmail());
     return new UserResponse(user.getUsername(), user.getEmail(), user.getZipCode(), user.getCity());
   }
 
@@ -77,7 +79,9 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     UserPrincipal userPrincipal;
     if (authentication.getPrincipal() instanceof UserDetails) {
       userPrincipal = (UserPrincipal) authentication.getPrincipal();
+      log.info("Successfully extracted UserPrincipal from authentication: {}", userPrincipal.getUsername());
     } else {
+      log.error("Authentication principal is not an instance of UserDetails. Authentication object: {}", authentication);
       throw new IllegalArgumentException("Authentication Invalid.");
     }
     long expireTime = getExpireTime(new Date());
@@ -85,10 +89,12 @@ public class AuthenticationServiceImpl implements AuthenticationService {
         userPrincipal,
         ACCESS_TOKEN,
         getExpireTime(new Date()));
+    log.info("Generated access token for user '{}': {}", userPrincipal.getUsername(), accessToken);
     String refreshToken = jwtService.generateRefreshToken(
         userPrincipal,
         TokenType.REFRESH_TOKEN,
         getExpireDay(new Date()));
+    log.info("Generated refresh token for user '{}': {}", userPrincipal.getUsername(), refreshToken);
     tokenService.save(new Token(accessToken, refreshToken, userPrincipal.getEmail()));
     return new AccessTokenResponse(accessToken, refreshToken, expireTime);
   }
@@ -96,25 +102,39 @@ public class AuthenticationServiceImpl implements AuthenticationService {
   @Override
   public AccessTokenResponse refresh(HttpServletRequest request) {
     String refreshToken = Optional.ofNullable(request.getHeader(REFRESH_TOKEN))
-        .orElseThrow(() -> new IllegalArgumentException("Refresh token is required."));
+        .orElseThrow(() -> {
+          log.error("Refresh token is missing in the request.");
+          return new IllegalArgumentException("Refresh token is required.");
+        });
     String email = jwtService.extractToken(refreshToken, TokenType.REFRESH_TOKEN);
+    log.info("Extracted email from refresh token: {}", email);
     User user = userRepository.findByEmail(email)
-        .orElseThrow(() -> new EmailNotFoundException("Not found email: " + email, HttpStatus.UNAUTHORIZED));
+        .orElseThrow(() -> {
+          log.error("User not found for email: {}", email);
+          return new EmailNotFoundException("Not found email: " + email, HttpStatus.UNAUTHORIZED);
+        });
     UserPrincipal userPrincipal = new UserPrincipal(user);
-    if (!(jwtService.isValidationToken(refreshToken, TokenType.REFRESH_TOKEN, userPrincipal))) {
+    if (!jwtService.isValidationToken(refreshToken, TokenType.REFRESH_TOKEN, userPrincipal)) {
+      log.error("Invalid refresh token provided for user: {}", email);
       throw new IllegalArgumentException("Refresh token invalid.");
     }
     long expireTime = getExpireTime(new Date());
     String accessToken = jwtService.generateToken(userPrincipal, ACCESS_TOKEN, expireTime);
     tokenService.save(new Token(accessToken, refreshToken, userPrincipal.getEmail()));
-    log.info("Token has been successfully saved.");
+    log.info("Access token generated and saved successfully for user: {}", email);
     return new AccessTokenResponse(accessToken, refreshToken, expireTime);
   }
 
   @Override
   public void logout(String accessToken) {
     String email = jwtService.extractToken(accessToken, ACCESS_TOKEN);
+    if (email == null) {
+      log.error("Failed to extract email from access token. Token may be invalid.");
+      throw new IllegalArgumentException("Invalid access token.");
+    }
+    log.info("Successfully extracted email {} from access token.", email);
     tokenService.delete(email);
+    log.info("User with email {} has been logged out and their token has been invalidated.", email);
   }
 
   private long getExpireTime(Date current) {
